+++ src/components/layout/Footer/index.tsx
@@ -1,4 +1,4 @@
-import styles from './css/Footer.module.css';
+import styles from './css/index.module.css';
 
 export default function Footer() {
   return (
